requirements.txt
Django>=1.8,<=1.8.99
djangorestframework==3.1.3
elasticsearch==2.4.1
requests==2.14.2
urllib3==1.21.1
django-localflavor==1.5